Refactor API into modular routers
