+++ api/app.py
@@ -567,280 +567,12 @@ def _prepare_messages(
     messages.append({"role": "user", "content": user_content})
     return messages, temperature
 
-# ---------- Endpoints ----------
-@app.get("/")
-def _root_to_app():
-    return RedirectResponse(url="/app/")
+# Include routers defined in api/routes
+from .routes.system import router as system_router
+from .routes.chat import router as chat_router
+from .routes.kb import router as kb_router
 
-@app.get("/health")
-def health(api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-    chroma_ok, cnt = True, None
-    try: cnt = collection.count()
-    except Exception: chroma_ok = False
-    # 簡易 Ollama 檢查
-    ollama_alive = False; ollama_models: List[str] = []
-    try:
-        import requests
-        rr = requests.get(f"{OLLAMA_HOST}/api/tags", timeout=3)
-        if rr.ok:
-            ollama_alive = True
-            jd = rr.json(); ollama_models = [t.get("name") for t in jd.get("models", [])]
-    except Exception: pass
-    return {
-        "status": "ok",
-        "db_path": DB_PATH,
-        "persist_dir": PERSIST_DIR,
-        "embedding_model": EMBEDDING_MODEL,
-        "collection": COLLECTION_NAME,
-        "docs_count": cnt, "chroma_ok": chroma_ok,
-        "language_default": DEFAULT_LANGUAGE,
-        "backends": {
-            "priority": MODEL_PRIORITY,
-            "openai": {"enabled": bool(OPENAI_API_KEY), "model": OPENAI_MODEL},
-            "ollama": {"enabled": bool(OLLAMA_MODEL), "model": OLLAMA_MODEL, "host": OLLAMA_HOST, "use_chat": OLLAMA_USE_CHAT, "alive": ollama_alive, "tags": ollama_models},
-        },
-    }
+app.include_router(system_router)
+app.include_router(chat_router)
+app.include_router(kb_router)
 
-@app.get("/threads")
-def list_threads(limit: int = 200, api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-    conn = _db()
-    cur = conn.execute("""
-        SELECT thread_id, MAX(ts) AS last_ts, COUNT(*) AS cnt
-        FROM messages GROUP BY thread_id ORDER BY last_ts DESC LIMIT ?
-    """, (int(limit),))
-    items = [{"thread_id": tid, "last_ts": ts, "count": cnt} for (tid, ts, cnt) in cur.fetchall()]
-    conn.close()
-    return {"threads": items}
-
-@app.get("/threads/{thread_id}/messages")
-def get_thread_messages(thread_id: str, limit: int = 500, api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-    conn = _db()
-    cur = conn.execute("SELECT ts, role, content, lang FROM messages WHERE thread_id=? ORDER BY id ASC LIMIT ?",
-                       (thread_id, int(limit)))
-    rows = cur.fetchall(); conn.close()
-    msgs = [{"ts": ts, "role": role, "content": content, "lang": lang} for (ts, role, content, lang) in rows]
-    return {"thread_id": thread_id, "messages": msgs, "summary": get_summary(thread_id)}
-
-@app.post("/docs/save")
-def docs_save(item: SaveDocItem, background_tasks: BackgroundTasks, api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-
-    Path(DOCS_DIR).mkdir(parents=True, exist_ok=True)
-    ts = datetime.utcnow().strftime("%Y%m%d_%H%M%S")
-    doc_id = f"doc-{ts}-{uuid.uuid4().hex[:8]}"
-    fname = f"{ts}_{_slug(item.title)}.json"
-
-    # NEW: 自動補 metadata 欄位
-    now_ts = int(time.time())
-    src_key = item.metadata.get("source_key") if isinstance(item.metadata, dict) else None
-    if not src_key:
-        src_key = _slug(item.title)  # 無則用標題 slug 當來源鍵
-    c_hash = content_hash(item.content)
-
-    metadata = dict(item.metadata or {})
-    metadata.update({
-        "updated_ts": now_ts,       # 供 recency 使用
-        "source_key": src_key,      # 供版本控制（唯一鍵）
-        "content_hash": c_hash,     # 供去重
-    })
-
-    payload = {
-        "id": doc_id,
-        "title": item.title,
-        "content": item.content,
-        "metadata": metadata,
-    }
-
-    (Path(DOCS_DIR) / fname).write_text(json.dumps(payload, ensure_ascii=False, indent=2), "utf-8")
-    background_tasks.add_task(_index_doc_to_stores, payload)  # 背景索引
-    return {"ok": True, "id": doc_id, "file": fname, "path": str(Path(DOCS_DIR) / fname)}
-
-@app.post("/kb/upsert")
-def kb_upsert(item: Dict[str, str], api_key: Optional[str] = Security(api_key_header)):
-    """最簡單的 KB upsert：{source, ref_id, title, content}"""
-    _auth(api_key)
-    for f in ("source","ref_id","title","content"):
-        if f not in item: raise HTTPException(400, f"missing field: {f}")
-    upsert_kb_item(item["source"], item["ref_id"], item["title"], item["content"])
-    return {"ok": True}
-
-@app.post("/kb/search")
-def kb_search(body: Dict[str, Any], api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-    q = (body.get("query") or "").strip()
-    if not q: raise HTTPException(400, "query required")
-    k = int(body.get("k") or 5)
-    return {"hits": search_kb_fts(q, limit=k)}
-
-# ========== /search：合併 Chroma + DB FTS 並過濾 ==========
-@app.post("/search")
-def search(req: SearchRequest, api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-    out = _query_chroma(req)
-    hits = out.get("hits", [])
-
-    try:
-        db_hits = search_kb_fts(req.query, limit=max(1, (req.k or 6)//2))
-        for h in db_hits:
-            h["metadata"] = {**h.get("metadata", {}), "title": h.get("title", ""), "source": "db"}
-
-        # 過濾 meta/schema 類（除非 canonicality=meta）
-        do_filter = FILTER_META_DEFAULT and not ((req.canonicality or "").strip().lower() == "meta")
-        hits = _filter_meta_hits(hits, do_filter) + _filter_meta_hits(db_hits, do_filter)
-
-        # 如果你有 recency/score 混合排序，這裡可再處理
-        # hits = _mix_with_recency(hits, alpha=0.7)
-
-    except Exception as e:
-        out["db_error"] = str(e)
-
-    # 可選：rerank / highlight
-    if req.rerank:
-        try:
-            hits = _rerank(req.query, hits); out["reranked"] = True
-        except Exception as e:
-            out["reranked"] = False; out["rerank_error"] = str(e)
-
-    if req.highlight:
-        for h in hits:
-            try:
-                h["highlights"] = _highlight(req.query, h.get("text", ""))
-            except Exception:
-                h["highlights"] = []
-
-    out["hits"] = hits[:max(1, req.k or 5)]
-    out["source"] = "chroma+db"
-    return out
-
-@app.post("/compose")
-def compose(req: ComposeRequest, api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-    # 檢索（合併文件＋DB）
-    hits = _search_internal(req.query, req.k, req.namespace, req.canonicality, req.rerank)
-    if not hits:
-        return {"draft": "", "citations": [], "note": "無檢索命中；請調整 query 或新增資料。"}
-    thread_id = req.thread_id or f"auto-{int(time.time()*1000)}"
-    lang = _norm_lang(req.language)
-    save_message(thread_id, "user", req.query, lang)
-
-    # 組歷史＋RAG
-    history_block = _format_history_block(thread_id, lang, max_turns=6, max_chars=1200)
-    rag_context, used_hits = _build_context_lang(hits, language=lang, max_chars=MAX_CONTEXT_CHARS, max_chars_per_chunk=600)
-    combined_ctx = (f"<<<HISTORY_START>>>\n{history_block}\n<<<HISTORY_END>>>\n\n" if history_block else "") + rag_context
-
-    messages, temperature = _prepare_messages(
-        req.query,
-        combined_ctx,
-        req.mode,
-        lang,
-        target_length=req.target_length,
-        style=req.style,
-    )
-    try:
-        draft, engine = generate(
-            messages,
-            engine=req.engine,
-            temperature=temperature,
-            max_tokens=req.max_tokens,
-            num_predict=req.num_predict,
-        )
-    except Exception as e:
-        raise HTTPException(500, f"LLM generate failed: {e}")
-
-    save_message(thread_id, "assistant", draft, lang)
-    prev = get_summary(thread_id)
-    update_src = f"Previous summary:\n{prev}\n\nNew exchange:\nU: {req.query}\nA: {draft}\n"
-    new_summary = _summarize_chunk_to_lang(update_src, lang, max_chars_per_chunk=800)
-    if new_summary: set_summary(thread_id, new_summary, lang)
-
-    cits = []
-    for h in hits:
-        m = h.get("metadata", {}) or {}
-        cits.append({"id": h.get("id"), "file_path": m.get("file_path"), "section": m.get("section")})
-    return {"draft": draft, "citations": cits, "used_hits": used_hits, "engine": engine, "language": lang, "thread_id": thread_id}
-
-@app.post("/compose_stream")
-def compose_stream(req: ComposeRequest, api_key: Optional[str] = Security(api_key_header)):
-    _auth(api_key)
-    headers = {
-        "Content-Type": "text/event-stream",
-        "Cache-Control": "no-cache",
-        "Connection": "keep-alive",
-        "X-Accel-Buffering": "no",
-    }
-    try:
-        hits = _search_internal(req.query, req.k, req.namespace, req.canonicality, req.rerank)
-        sel = getattr(req, "selected_ids", None)
-        if sel:
-            idset = set(sel)
-            hits = [h for h in hits if h.get("id") in idset]
-
-        thread_id = req.thread_id or f"auto-{int(time.time()*1000)}"
-        lang = _norm_lang(req.language)
-        save_message(thread_id, "user", req.query, lang)
-        debug_prompts = bool(getattr(req, "debug", False) or os.environ.get("DEBUG_PROMPTS") in ("1","true","True"))
-
-        history_block = _format_history_block(thread_id, lang, max_turns=6, max_chars=1200)
-        rag_context, used_hits = _build_context_lang(hits, language=lang, max_chars=MAX_CONTEXT_CHARS, max_chars_per_chunk=600)
-        combined_ctx = (f"<<<HISTORY_START>>>\n{history_block}\n<<<HISTORY_END>>>\n\n" if history_block else "") + rag_context
-    except Exception as e:
-        # 把初始化錯誤改成 SSE 回給前端，而不是 500
-        def init_fail():
-            yield "data: " + json.dumps({"token": f"[compose_stream init error] {e}"}, ensure_ascii=False) + "\n\n"
-        return StreamingResponse(init_fail(), media_type="text/event-stream", headers=headers)
-
-    def event_stream():
-        assistant_accum: List[str] = []
-        try:
-            messages, temperature = _prepare_messages(
-                req.query,
-                combined_ctx,
-                req.mode,
-                lang,
-                target_length=req.target_length,
-                style=req.style,
-            )
-            yield f"data: {json.dumps({'token': ''})}\n\n"
-            stream, final_engine = generate(
-                messages,
-                engine=req.engine,
-                stream=True,
-                temperature=temperature,
-                max_tokens=req.max_tokens,
-                num_predict=req.num_predict,
-            )
-            if debug_prompts:
-                model = OPENAI_MODEL if final_engine == 'openai' else OLLAMA_MODEL
-                preview = _preview_messages(final_engine, model, messages)
-                yield "data: " + json.dumps({"debug": preview}, ensure_ascii=False) + "\n\n"
-            for token in stream:
-                assistant_accum.append(token)
-                yield f"data: {json.dumps({'token': token})}\n\n"
-        except Exception as e:
-            yield f'data: {json.dumps({"token": f"[compose_stream error] {e}"})}\n\n'
-            return
-
-        try:
-            assistant_text = "".join(assistant_accum).strip()
-            if assistant_text:
-                save_message(thread_id, "assistant", assistant_text, lang)
-                prev = get_summary(thread_id)
-                update_src = f"Previous summary:\n{prev}\n\nNew exchange:\nU: {req.query}\nA: {assistant_text}\n"
-                new_summary = _summarize_chunk_to_lang(update_src, lang, max_chars_per_chunk=800)
-                if new_summary:
-                    set_summary(thread_id, new_summary, lang)
-
-            tail = {
-                "citations": hits,
-                "used_hits": used_hits,
-                "engine": final_engine,
-                "thread_id": thread_id,
-            }
-            yield f"data: {json.dumps(tail, ensure_ascii=False)}\n\n"
-        except Exception as e:
-            yield f'data: {json.dumps({"token": f"[compose_stream error] {e}"})}\n\n'
-
-    return StreamingResponse(event_stream(), media_type="text/event-stream", headers=headers)
